+++ BloqueRoca.h
@@ -14,7 +14,7 @@ public:
 	BloqueRoca(int identidad, Mundo& mundo, const b2Vec2& posicion);
 	BloqueRoca(BloqueRoca&& otro);
 	void recibirDisparo(Disparo* disparo);
-	int getId();
+	uint8_t getId();
 	const b2Vec2& getPosition();
 	void empezarContacto(Cuerpo* otro);
 };
